fix: Leader hotkeys not working in 2nd+ window
This commit contains a fix for an issue where Leader Hotkeys is not able
to respond to key sequences in extra windows created by obsidian.

As of this commit, there is no official API for iterating Obsidian windows.
This feature relies on iterating an undocumented field inside the
`app.workspace` object.